+++ lib/preferences.dart
@@ -18,12 +18,15 @@ const String INV_HOME_SHOW_SUPPLIERS = "homeShowSuppliers";
 const String INV_SOUNDS_BARCODE = "barcodeSounds";
 const String INV_SOUNDS_SERVER = "serverSounds";
 
+// Part settings
+const String INV_PART_SHOW_PARAMETERS = "partShowParameters";
+const String INV_PART_SHOW_BOM = "partShowBom";
+
+// Stock settings
 const String INV_STOCK_SHOW_HISTORY = "stockShowHistory";
 
 const String INV_REPORT_ERRORS = "reportErrors";
-
 const String INV_STRICT_HTTPS = "strictHttps";
-
 
 /*
  * Class for storing InvenTree preferences in a NoSql DB