+++ assets/release_notes.md
@@ -6,6 +6,7 @@
 
 - Add support for structural part categories
 - Add support for structural stock locations
+- Allow deletion of attachments via app
 - Updated translations