+++ assets/release_notes.md
@@ -8,6 +8,7 @@
 - Add support for structural part categories
 - Add support for structural stock locations
 - Allow deletion of attachments via app
+- Adds option for controlling BOM display
 - Updated translations